+++ packages/terra-dynamic-grid/CHANGELOG.md
@@ -1,6 +1,11 @@
 ChangeLog
 =========
 
-1.0.0 - (<!--- Insert date using June 28, 2017 format. --->)
+Unreleased
 -----------------
+### Changed
+* Update nightwatch tests and test scripts.
+
+1.0.0 - (August 31, 2017)
+----------------
 Initial stable release