+++ packages/terra-embedded-content-consumer/CHANGELOG.md
@@ -2,4 +2,10 @@ ChangeLog
 =========
 
 Unreleased
-----------
+-----------------
+### Changed
+* Update nightwatch tests and test scripts.
+
+1.0.0 - (August 31, 2017)
+----------------
+Initial stable release